Continua el desarrollo de la asignacion de profesores

+++ application/models/pride/evaluadorevaluado.php
@@ -37,9 +37,13 @@ class EvaluadorEvaluado extends ActiveRecord\Model{
 	function desasignar($idEvaluador,$idEvaluado) {
 	
 	
-		$evaluador_evaluado = evaluador_evaluado::first(array("conditions" => array("id_evaluador = ? AND id_evaluado = ?",$idEvaluador,$idEvaluado) ));
-		$evaluador_evaluado->delete();
-	
+		$evaluador_evaluado = EvaluadorEvaluado::first(array("conditions" => array("id_evaluador = ? AND id_evaluado = ?",$idEvaluador,$idEvaluado) ));
+		if(sizeof($evaluador_evaluado)){
+			$evaluador_evaluado->delete();
+			$respuesta["respuesta"] = array("mensaje" => "Profesor desasignado");
+		}
+		
+		echo json_encode($respuesta);
 	
 	}
